+++ services/businessLayer/teamPlanHierarchy/refactorL5L6_level5Lib.xsjslib
@@ -152,35 +152,43 @@ function getHl5ById(hl5Id, carryOver) {
         throw ErrorLib.getErrors().BadRequest("The Parameter ID is not found", "hl5Services/handleGet/getHl5ById", L5_MSG_INITIATIVE_NOT_FOUND);
 
     var hl5 = JSON.parse(JSON.stringify(dataHl5.getHl5ById(hl5Id)));
-    var internalCofunding = getInternalCofunding(hl5Id);
-    var externalCofunding = getExternalCofunding(hl5Id);
     var targetKpi = expectedOutcomesLib.getExpectedOutcomesByHl5Id(hl5Id, hl5.HL4_ID);
-
     hl5.TARGET_KPIS = carryOver ? expectedOutcomesLib.filterKpiByLevel(targetKpi, 'HL6') : targetKpi;
-    hl5.BUDGET = Number(hl5.BUDGET);
-    hl5.PARTNERS = externalCofunding.PARTNERS;
-    hl5.INTEL_TOTAL_BUDGET = externalCofunding.PARTNER_INTEL_TOTAL;
-    hl5.PARTNER_CURRENCY_ID = externalCofunding.PARTNER_CURRENCY_ID;
-    hl5.EXTERNAL_TOTAL_BUDGET = externalCofunding.PARTNER_EXTERNAL_TOTAL;
 
-    hl5.BUDGET_DISTRIBUTION = dataHl5.getHl5MyBudgetByHl5Id(hl5Id);
+    if (!carryOver) {
+        var internalCofunding = getInternalCofunding(hl5Id);
+        var externalCofunding = getExternalCofunding(hl5Id);
+        hl5.BUDGET = Number(hl5.BUDGET);
+        hl5.PARTNERS = externalCofunding.PARTNERS;
+        hl5.INTEL_TOTAL_BUDGET = externalCofunding.PARTNER_INTEL_TOTAL;
+        hl5.PARTNER_CURRENCY_ID = externalCofunding.PARTNER_CURRENCY_ID;
+        hl5.EXTERNAL_TOTAL_BUDGET = externalCofunding.PARTNER_EXTERNAL_TOTAL;
 
-    hl5.SALES = internalCofunding.SALE;
-    hl5.SALE_TOTAL = internalCofunding.SALE_TOTAL;
-    hl5.SALE_CURRENCY_ID = internalCofunding.SALE_CURRENCY_ID;
-    hl5.SALE_REQUESTS = internalCofunding.SALE_REQUESTS;
+        hl5.BUDGET_DISTRIBUTION = dataHl5.getHl5MyBudgetByHl5Id(hl5Id);
 
-    hl5.CATEGORIES = carryOver ? level6Lib.getCarryOverHl5CategoryOption(hl5Id) : getCategoryOption(hl5Id);
-    hl5.SERVICE_REQUEST_CATEGORIES = getServiceRequestCategoryOptionByHl5Id(hl5Id);
+        hl5.SALES = internalCofunding.SALE;
+        hl5.SALE_TOTAL = internalCofunding.SALE_TOTAL;
+        hl5.SALE_CURRENCY_ID = internalCofunding.SALE_CURRENCY_ID;
+        hl5.SALE_REQUESTS = internalCofunding.SALE_REQUESTS;
 
-    hl5.TOTAL_BUDGET = (hl5.BUDGET +
-        externalCofunding.PARTNER_INTEL_TOTAL / externalCofunding.PARTNER_CURRENCY_VALUE +
-        hl5.SALE_TOTAL / internalCofunding.SALE_CURRENCY_VALUE +
-        externalCofunding.PARTNER_EXTERNAL_TOTAL / externalCofunding.PARTNER_CURRENCY_VALUE).toFixed(2);
+        hl5.CATEGORIES = getCategoryOption(hl5Id);
+        hl5.SERVICE_REQUEST_CATEGORIES = getServiceRequestCategoryOptionByHl5Id(hl5Id);
 
-    hl5.IS_IN_CRM = !!dataHl5.hl5ExistsInCrm(hl5Id);
+        hl5.TOTAL_BUDGET = (hl5.BUDGET +
+            externalCofunding.PARTNER_INTEL_TOTAL / externalCofunding.PARTNER_CURRENCY_VALUE +
+            hl5.SALE_TOTAL / internalCofunding.SALE_CURRENCY_VALUE +
+            externalCofunding.PARTNER_EXTERNAL_TOTAL / externalCofunding.PARTNER_CURRENCY_VALUE).toFixed(2);
 
-    hl5.BUDGET = hl5.BUDGET * Number(hl5.CURRENCY_VALUE);
+        hl5.BUDGET = hl5.BUDGET * Number(hl5.CURRENCY_VALUE);
+        hl5.IS_IN_CRM = !!dataHl5.hl5ExistsInCrm(hl5Id);
+    } else {
+        hl5.BUDGET = 0;
+        hl5.CATEGORIES = level6Lib.getCarryOverHl5CategoryOption(hl5Id);
+        hl5.HL5_CRM_DESCRIPTION = undefined;
+        hl5.ACRONYM = undefined;
+        hl5.CO_FUNDED = 0;
+        hl5.ALLOW_BUDGET_ZERO = 0;
+    }
 
     return serverToUiParser(hl5);
 }
@@ -344,7 +352,7 @@ function insertHl5RequestCategoryOption(hl5Id, requestCategoryOption, userId) {
 
             if (option.FORM && option.FORM_NAME) {
                 option.FORM.HL5_ID = hl5Id;
-                option.FORM.FORM_STATUS_DETAIL_ID = option.FORM_STATUS_DETAIL_ID;
+                // option.FORM.FORM_STATUS_DETAIL_ID = option.FORM_STATUS_DETAIL_ID;
                 switch (option.FORM_NAME) {
                     case 'SegmentationForm':
                         blSegmentation.processForm(option.FORM, userId);
@@ -633,7 +641,7 @@ function updateHl5(data, userId) {
         setHl5Status(data.HL5_ID, data.STATUS_DETAIL_ID, userId);
         var ownMoneyBudgetSpendRequestStatus = budgetSpendRequest.getOwnMoneyBudgetSpendRequestStatusByHlIdLevel(hl5_id, 'HL5');
         var automaticBudgetApproval = blLevel2.getHl2AllowAutomaticBudgetApprovalByHl4Id(data.HL4_ID) && data.IN_BUDGET && !!Number(data.BUDGET);
-        if(!ownMoneyBudgetSpendRequestStatus || ownMoneyBudgetSpendRequestStatus == budgetSpendRequestStatus.NO_LONGER_REQUESTED){
+        if (!ownMoneyBudgetSpendRequestStatus || ownMoneyBudgetSpendRequestStatus == budgetSpendRequestStatus.NO_LONGER_REQUESTED) {
             budgetSpendRequest.insertOwnMoneyBudgetSpendRequest(data.BUDGET, hl5_id, 'HL5', userId, automaticBudgetApproval);
         } else {
             if (objHL5.BUDGET != data.BUDGET) {
@@ -1344,12 +1352,15 @@ function checkPermission(userSessionID, method, hl5Id) {
 }
 
 function findHLSalesId(Sales, OrganizationId, OrganizationType) {
+    var hlSaleId = null;
     for (var i = 0; i < Sales.length; i++) {
         if (Sales[i].ORGANIZATION_ID == OrganizationId && Sales[i].ORGANIZATION_TYPE == OrganizationType) {
-            return Sales[i].HL5_SALE_ID;
+            hlSaleId = Sales[i].HL5_SALES_ID;
+            break;
         }
     }
-    return null;
+
+    return hlSaleId;
 }
 
 function getNewSerialAcronym(Hl4_id) {
@@ -1379,7 +1390,7 @@ function getNewSerialAcronym(Hl4_id) {
             }
         }
 
-    	if (newSerial > 99) {
+        if (newSerial > 99) {
             throw ErrorLib.getErrors().CustomError("", "", "It is not possible to create more Marketing Tactics for this Marketing Campaign.");
         }
     }
@@ -1575,7 +1586,7 @@ function insertExternalCoFunding(data, userId) {
                     ,
                     in_claim_id: PARTNER_TYPE.INTEL == partner.PARTNER_TYPE_ID ? partner.CLAIM_ID : null
                     ,
-                    in_comments: PARTNER_TYPE.INTEL == partner.PARTNER_TYPE_ID ? partner.NOTES : null
+                    in_comments: PARTNER_TYPE.INTEL == partner.PARTNER_TYPE_ID ? partner.COMMENTS : null
                     ,
                     in_company_name: PARTNER_TYPE.EXTERNAL_PARTNER == partner.PARTNER_TYPE_ID ? partner.COMPANY_NAME : null
                     ,
@@ -1598,9 +1609,9 @@ function updateExternalCoFunding(data, userId) {
         var externalCoFundingCurrency = dataCurrency.getCurrencyValueId(data.PARTNER_CURRENCY_ID);
         data.PARTNERS.forEach(function (partner) {
             if (!partner.PARTNER_ID) {
-                var budgetSpendRequestId = budgetSpendRequest.insertPartnerBudgetSpendRequest(partner.AMOUNT, partner.MESSAGE, hl5_id, 'HL5', externalCoFundingCurrency, userId);
+                var budgetSpendRequestId = budgetSpendRequest.insertPartnerBudgetSpendRequest(partner.AMOUNT, partner.MESSAGE, data.HL5_ID, 'HL5', externalCoFundingCurrency, userId);
                 arrPartnerToInsert.push({
-                    in_hl5_id: hl5_id
+                    in_hl5_id: data.HL5_ID
                     ,
                     in_partner_name: null
                     ,
@@ -1620,7 +1631,7 @@ function updateExternalCoFunding(data, userId) {
                     ,
                     in_claim_id: PARTNER_TYPE.INTEL == partner.PARTNER_TYPE_ID ? partner.CLAIM_ID : null
                     ,
-                    in_comments: PARTNER_TYPE.INTEL == partner.PARTNER_TYPE_ID ? partner.NOTES : null
+                    in_comments: PARTNER_TYPE.INTEL == partner.PARTNER_TYPE_ID ? partner.COMMENTS : null
                     ,
                     in_company_name: PARTNER_TYPE.EXTERNAL_PARTNER == partner.PARTNER_TYPE_ID ? partner.COMPANY_NAME : null
                     ,
@@ -1656,7 +1667,7 @@ function updateExternalCoFunding(data, userId) {
                     ,
                     in_claim_id: PARTNER_TYPE.INTEL == partner.PARTNER_TYPE_ID ? partner.CLAIM_ID : null
                     ,
-                    in_comments: PARTNER_TYPE.INTEL == partner.PARTNER_TYPE_ID ? partner.NOTES : null
+                    in_comments: PARTNER_TYPE.INTEL == partner.PARTNER_TYPE_ID ? partner.COMMENTS : null
                     ,
                     in_company_name: PARTNER_TYPE.EXTERNAL_PARTNER == partner.PARTNER_TYPE_ID ? partner.COMPANY_NAME : null
                     ,
@@ -1677,6 +1688,7 @@ function updateExternalCoFunding(data, userId) {
         if (arrPartnerToUpdate.length) {
             dataPartner.updatePartner(arrPartnerToUpdate, 'HL5');
         }
+
         if (budgetSpendRequestToUpdate.length)
             budgetSpendRequest.updateBudgetSpendRequest(budgetSpendRequestToUpdate, userId, true);
     }
@@ -1686,8 +1698,9 @@ function updateExternalCoFunding(data, userId) {
         var arrPartnerToDelete = data.PARTNERS_REQUESTS_REMOVED.map(function (id) {
             return {in_partner_id: id, in_user_id: userId};
         });
-        arrPartnerToDelete = [];
         var pendingPartner = dataPartner.getPendingPartnerByPartnerId(arrPartnerToDelete, "HL5");
+        arrPartnerToDelete = [];
+
         pendingPartner.forEach(function (elem) {
             arrPartnerToDelete.push({in_partner_id: elem.IN_PARTNER_ID, in_user_id: userId});
             arrBudgetSpendRequestToDelete.push({
@@ -1695,9 +1708,10 @@ function updateExternalCoFunding(data, userId) {
                 in_user_id: userId
             });
         });
-
-        dataPartner.deleteHlPartnerByPartnerId(arrPartnerToDelete, "HL5");
-        budgetSpendRequest.deleteBudgetSpendRequest(arrBudgetSpendRequestToDelete);
+        if (arrPartnerToDelete && arrPartnerToDelete.length) {
+            dataPartner.deleteHlPartnerByPartnerId(arrPartnerToDelete, "HL5");
+            budgetSpendRequest.deleteBudgetSpendRequest(arrBudgetSpendRequestToDelete);
+        }
     }
     return true;
 }
@@ -1876,42 +1890,59 @@ function serverToUiParser(object) {
         REGIONS: [],
         CENTRAL_TEAMS: []
     };
-
-    object.BUDGET_DISTRIBUTION.forEach(function (obj) {
-        var aux = {};
-        if (obj.ORGANIZATION_TYPE === 1) {
-            aux.REGION_ID = obj.ORGANIZATION_ID;
-            aux.PERCENTAGE = obj.PERCENTAGE;
-            aux.REGION_NAME = obj.ORGANIZATION_NAME;
-            BUDGET_DISTRIBUTION.REGIONS.push(aux);
-        } else {
-            aux.HL2_ID = obj.ORGANIZATION_ID;
-            aux.PERCENTAGE = obj.PERCENTAGE;
-            aux.ACRONYM = obj.ORGANIZATION_NAME;
-            BUDGET_DISTRIBUTION.CENTRAL_TEAMS.push(aux);
-        }
-    });
-    object.SALES.forEach(function (obj) {
-        var aux = {};
-        aux.HL5_SALES_ID = obj.HL5_SALES_ID;
-        if (obj.ORGANIZATION_TYPE === 1) {
-            aux.REGION_ID = obj.ORGANIZATION_ID;
-            aux.AMOUNT = obj.AMOUNT;
-            aux.REGION_NAME = obj.ORGANIZATION_NAME;
-            SALE.REGIONS.push(aux);
-        } else if (obj.ORGANIZATION_TYPE === 2) {
-            aux.HL2_ID = obj.ORGANIZATION_ID;
-            aux.AMOUNT = obj.AMOUNT;
-            aux.ACRONYM = obj.ORGANIZATION_NAME;
-            SALE.CENTRAL_TEAMS.push(aux);
-        } else {
-            aux.DESCRIPTION = obj.DESCRIPTION;
-            aux.AMOUNT = obj.AMOUNT;
-            SALE.OTHERS.push(aux);
-        }
-        aux.MESSAGE = obj.MESSAGE;
-    });
-
+    if(object.BUDGET_DISTRIBUTION) {
+        object.BUDGET_DISTRIBUTION.forEach(function (obj) {
+            var aux = {};
+            if (obj.ORGANIZATION_TYPE === 1) {
+                aux.REGION_ID = obj.ORGANIZATION_ID;
+                aux.PERCENTAGE = obj.PERCENTAGE;
+                aux.REGION_NAME = obj.ORGANIZATION_NAME;
+                BUDGET_DISTRIBUTION.REGIONS.push(aux);
+            } else {
+                aux.HL2_ID = obj.ORGANIZATION_ID;
+                aux.PERCENTAGE = obj.PERCENTAGE;
+                aux.ORGANIZATION_ACRONYM = obj.ORGANIZATION_NAME;
+                BUDGET_DISTRIBUTION.CENTRAL_TEAMS.push(aux);
+            }
+        });
+    }
+    if(object.SALES) {
+        object.SALES.forEach(function (obj) {
+            var aux = {};
+            aux.SALES_ID = obj.HL5_SALES_ID;
+            if (obj.ORGANIZATION_TYPE === 1) {
+                aux.REGION_ID = obj.ORGANIZATION_ID;
+                aux.AMOUNT = obj.AMOUNT;
+                aux.REGION_NAME = obj.ORGANIZATION_NAME;
+                SALE.REGIONS.push(aux);
+            } else if (obj.ORGANIZATION_TYPE === 2) {
+                aux.HL2_ID = obj.ORGANIZATION_ID;
+                aux.AMOUNT = obj.AMOUNT;
+                aux.ORGANIZATION_ACRONYM = obj.ORGANIZATION_NAME;
+                SALE.CENTRAL_TEAMS.push(aux);
+            } else {
+                aux.DESCRIPTION = obj.DESCRIPTION;
+                aux.AMOUNT = obj.AMOUNT;
+                SALE.OTHERS.push(aux);
+            }
+            aux.MESSAGE = obj.MESSAGE;
+        });
+    }
+    if(object.SALE_REQUESTS){
+        object.SALE_REQUESTS.forEach(function (obj) {
+            var attribute = obj.ORGANIZATION_TYPE == ORGANIZATION_TYPE.REGIONAL ? 'REGION_ID'
+                : obj.ORGANIZATION_TYPE == ORGANIZATION_TYPE.CENTRAL ? 'HL2_ID'
+                    : 'ORGANIZATION_ID';
+            obj[attribute] = obj.ORGANIZATION_ID;
+            obj.ORGANIZATION_ID = undefined;
+        });
+    }
+    if(object.PARTNERS) {
+        object.PARTNERS.forEach(function (obj) {
+            obj.MESSAGE = obj.DESCRIPTION;
+            obj.DESCRIPTION = undefined;
+        });
+    }
     object.SALES = SALE;
     object.BUDGET_DISTRIBUTION = BUDGET_DISTRIBUTION;
     object.CRM_DESCRIPTION = object.HL5_CRM_DESCRIPTION;
